chore(js): 千星奇域适配「月之四」版本
