+++ repo/js/MiliastraExperiencePlayback/assets/playbacks/美景音乐播放器结算.json
@@ -20,7 +20,7 @@
     { "type": 1, "keyCode": 68, "mouseX": 0, "mouseY": 0, "time": 10718 },
     { "type": 0, "keyCode": 65, "mouseX": 0, "mouseY": 0, "time": 11047 },
     { "type": 1, "keyCode": 65, "mouseX": 0, "mouseY": 0, "time": 11250 },
-    { "type": 1, "keyCode": 87, "mouseX": 0, "mouseY": 0, "time": 11703 },
+    { "type": 1, "keyCode": 87, "mouseX": 0, "mouseY": 0, "time": 11800 },
 
     { "type": 0, "keyCode": 70, "mouseX": 0, "mouseY": 0, "time": 12562 },
     { "type": 1, "keyCode": 70, "mouseX": 0, "mouseY": 0, "time": 12718 },
