+++ repo/js/MiliastraExperiencePlayback/libs/rolldown-runtime.js
@@ -1,0 +1,10 @@
+//#region rolldown:runtime
+var __defProp = Object.defineProperty;
+var __name = (target, value) =>
+  __defProp(target, "name", {
+    value,
+    configurable: true,
+  });
+
+//#endregion
+export { __name };
